# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Shortener",
        "description": "Listar los shorteners de un usuario",
        "version": "1.0"
    },
    "paths": {
        "/api/{user}": {
            "get": {
                "tags": [
                    "Shortener"
                ],
                "summary": "Listar los shorteners de un usuario",
                "operationId": "cd232f4ef82302e8610f516ae6acee0a",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Id del usuario",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listar los shorteners de un usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "25680da0-036b-4e9e-bc33-8f74495c853b"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://imgur.com/gallery/uh-i-don-t-remember-saving-this-dSepCYe"
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "wrtvZhHs"
                                                    },
                                                    "vists": {
                                                        "type": "number",
                                                        "example": "15"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2024-02-23T00:09:16.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "next": {
                                                        "type": "string|NULL",
                                                        "example": "http://localhost:8000/api/shorten/wrtvZhHs?page=2"
                                                    },
                                                    "prev": {
                                                        "type": "string|NULL",
                                                        "example": "http://localhost:8000/api/shorten/wrtvZhHs?page=1"
                                                    },
                                                    "total": {
                                                        "type": "number",
                                                        "example": "30"
                                                    },
                                                    "per_page": {
                                                        "type": "number",
                                                        "example": "10"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/shorten/{code}": {
            "get": {
                "tags": [
                    "Shortener"
                ],
                "summary": "Obtener la url original de un shortener",
                "operationId": "17c5bd937a8978bf44d81e066f33c750",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "codigo del shortener",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener la url original de un shortener",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "url": {
                                            "type": "string",
                                            "example": "https://imgur.com/gallery/uh-i-don-t-remember-saving-this-dSepCYe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/shorten": {
            "post": {
                "tags": [
                    "Shortener"
                ],
                "summary": "Crear un shortener",
                "operationId": "ab67d30c33fa486fcedac58366c05af1",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "example": "https://imgur.com/gallery/uh-i-don-t-remember-saving-this-dSepCYe"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "example": "25680da0-036b-4e9e-bc33-8f74495c853b"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crear un shortener",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://imgur.com/gallery/uh-i-don-t-remember-saving-this-dSepCYe"
                                                    },
                                                    "code": {
                                                        "type": "string",
                                                        "example": "dSepCYeH"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Shortener",
            "description": "Shortener"
        }
    ]
}